back to double, cause of some strange behaviour

+++ RPOSC-LogicAnalyser/src/Acquirer.hpp
@@ -7,5 +7,5 @@
 class Acquirer{
     public:
         virtual bool startAcq(){};
-        virtual vector<float> getData(int channel){};
+        virtual vector<double> getData(int channel){};
 };
